refactor: 회원가입 정보입력창을 joinInputBox로 분리
