+++ src/types/types.ts
@@ -1,0 +1,7 @@
+export interface userInfo {
+  loginId: string;
+  password: string;
+  realName: string;
+  phone: string;
+  birthDate: string;
+}
